Move ensemble functions - add tests
