+++ xclim/ensembles/__init__.py
@@ -1,0 +1,4 @@
+"""Ensembles Module."""
+from .base import create_ensemble, ensemble_mean_std_max_min, ensemble_percentiles
+from .reduce import kkz_reduce_ensemble, kmeans_reduce_ensemble, plot_rsqprofile
+from .robustness import ensemble_robustness
